Fix SIGCHLD being masked for child processes on Linux
By default, children created with posix_spawn inherit the
signal mask from their parent.  On Linux, B was masking
SIGCHLD.  This caused SIGCHLD to be masked in the child,
leading to problems when invoking e.g. the GCC driver.

This patch tells posix_spawn to create the child with an
empty signal mask.