Mark default cases after switch unreachable
Some compilers (such as GCC 4.9) warn about no returns in
case the value of an enum type isn't one of the values
defined for that enum. Mark such cases as unreachable.

In the future, adding -Wswitch-default -Wswitch-enum may be
valuable for making this systematic and not relevant to
escape analysis.