Implement Python JoinFiles example
Implement:

* Python wrapper classes for B's native classes, as a native
  module (compiled using CMake).
* An example program emulating the pre-existing C JoinFiles
  example.
* Some basic unit tests to catch crashes.

The implementation is just enough to support the example.
Many pieces are missing, such as implementing RunLoop-s and
Question classes in Python and the Question class methods.

I am new familiar with the CPython C API, so I apologize for
hilarious code. /o\ Implementation is subject to change,
naturally, as I learn the API.